no need for 64bit variables

+++ packet_manip.h
@@ -17,6 +17,6 @@ int combine_packets( struct qserver *server );
 int add_packet( struct qserver *server, unsigned int pkt_id, int pkt_index, int pkt_max, int datalen, char *data, int calc_max );
 int next_sequence();
 SavedData* get_packet_fragment( int index );
-unsigned long combined_length( struct qserver *server, int pkt_id );
+unsigned combined_length( struct qserver *server, int pkt_id );
 
 #endif
